SNOW-1673337: Move indexing.py into src/…/modin/plugin/extensions/
