+++ tests/unit/modin/test_utils.py
@@ -14,7 +14,6 @@ from snowflake.snowpark import Column
 from snowflake.snowpark._internal.analyzer.analyzer_utils import DOUBLE_QUOTE
 from snowflake.snowpark._internal.type_utils import VALID_PYTHON_TYPES_FOR_LITERAL_VALUE
 from snowflake.snowpark.functions import col
-from snowflake.snowpark.modin.pandas.indexing import is_boolean_array
 from snowflake.snowpark.modin.plugin._internal.ordered_dataframe import OrderedDataFrame
 from snowflake.snowpark.modin.plugin._internal.utils import (
     _MAX_IDENTIFIER_LENGTH,
@@ -47,6 +46,9 @@ from snowflake.snowpark.modin.plugin._internal.utils import (
     to_pandas_label,
     try_convert_to_simple_slice,
     unquote_name_if_quoted,
+)
+from snowflake.snowpark.modin.plugin.extensions.indexing_overrides import (
+    is_boolean_array,
 )
